Improvements to array-of-u8 handling in tor-bytes.
Now that we have const generics, we can use them.  We can also avoid
an extra clone in the implementation for [u8; N].

Nothing in our codebase requires that we use Reader or Writer on a
GenericArray holding anything other than u8, so I've switched back
to the more efficient implementation there.

I've added a fuzzer case for the new method, but apparently rustc nightly isn't working too
well with fuzzers for me; I'm going to try it tomorrow.