tor-bytes: Split EncodeError from Error
Since there is currently only one error type that can occur when
encoding, it doesn't make sense to use the full Error type here.

This split will help us downstream, as we no longer need to
categorize tor_bytes::Error as "an error in encoding or decoding".

I considered renaming Error to DecodeError, but that had pretty huge
downstream effects, and didn't seem to be worth it.